chore: use drawer to display table detail
