+++ frontend/src/components/TableTable.vue
@@ -56,21 +56,30 @@
       </tfoot>
     </template>
   </BBTable>
+
+  <TableDetailDrawer
+    v-if="state.selectedTableName"
+    :database-name="database.name"
+    :schema-name="schemaName"
+    :table-name="state.selectedTableName"
+    @dismiss="state.selectedTableName = undefined"
+  />
 </template>
 
 <script lang="ts" setup>
 import { computed, PropType, reactive } from "vue";
-import { useRouter } from "vue-router";
 import { useI18n } from "vue-i18n";
 import { ComposedDatabase } from "@/types";
 import { TableMetadata } from "@/types/proto/store/database";
-import { bytesToString, databaseV1Slug, isGhostTable } from "@/utils";
+import { bytesToString, isGhostTable } from "@/utils";
 import EllipsisText from "@/components/EllipsisText.vue";
 import { Engine } from "@/types/proto/v1/common";
 import { BBTableColumn } from "@/bbkit";
+import TableDetailDrawer from "./TableDetailDrawer.vue";
 
 type LocalState = {
   showReservedTableList: boolean;
+  selectedTableName?: string;
 };
 
 const props = defineProps({
@@ -88,7 +97,6 @@ const props = defineProps({
   },
 });
 
-const router = useRouter();
 const { t } = useI18n();
 
 const state = reactive<LocalState>({
@@ -166,17 +174,6 @@ const mixedTableList = computed(() => {
 });
 
 const clickTable = (_: number, row: number, e: MouseEvent) => {
-  const table = mixedTableList.value[row];
-  let url = `/db/${databaseV1Slug(props.database)}/table/${encodeURIComponent(
-    table.name
-  )}`;
-  if (props.schemaName !== "") {
-    url = url + `?schema=${encodeURIComponent(props.schemaName)}`;
-  }
-  if (e.ctrlKey || e.metaKey) {
-    window.open(url, "_blank");
-  } else {
-    router.push(url);
-  }
+  state.selectedTableName = mixedTableList.value[row].name;
 };
 </script>
